Track which cards aren't part of any set and return them as results
This is a step towards being able to play through the whole deck, as
the actual game is played. Next we need to try to choose some, hopefully
optimal subset of all the sets on the table to pick up, leaving the
remaining cards, so we can deal more and go to the next round.